feat: Complete phases 8-12 - WebSocket, Job Queue, Config Rollback, Systemd, CI/CD
- Phase 8: WebSocket real-time communication with hub/client architecture
- Phase 9: Job queue system with asynq (backup, restore, deploy, SSL, cleanup)
- Phase 10: Config rollback with snapshots, diff, templates
- Phase 11: Node.js app systemd integration for process management
- Phase 12: CI/CD pipeline with GitHub Actions and deployment scripts

Also fixed:
- Go compilation errors (duplicate types, missing imports, method signatures)
- Frontend build errors (missing UI components, TypeScript types)
- Added UI component library (Card, Button, Badge, Input, Select, Tabs)
- Added deployment systemd service files and scripts

+++ backend/internal/handler/website.go
@@ -64,7 +64,7 @@ func (h *WebsiteHandler) List(c *gin.Context) {
 		return
 	}
 
-	utils.Paginated(c, websites, total, params.Page, params.PerPage)
+	utils.Paginated(c, websites, int64(total), params.Page, params.PerPage)
 }
 
 func (h *WebsiteHandler) Update(c *gin.Context) {
